рефакторинг отправки сигнала, исправление warnings

+++ kernel/proc/syscalls.h
@@ -11,8 +11,8 @@
 #include "kairax/poll.h"
 
 #define VALIDATE_NULL_POINTER(base) if (base == 0) {return -EFAULT;}
-#define VALIDATE_USER_POINTER(proc,base,len) if (process_is_userspace_region(proc,base,len,0) == 0) {return -EFAULT;}
-#define VALIDATE_USER_POINTER_PROTECTION(proc,base,len,prot) if (process_is_userspace_region(proc,base,len,prot) == 0) {return -EFAULT;}
+#define VALIDATE_USER_POINTER(proc,base,len) if (process_is_userspace_region(proc,(uintptr_t)base,len,0) == 0) {return -EFAULT;}
+#define VALIDATE_USER_POINTER_PROTECTION(proc,base,len,prot) if (process_is_userspace_region(proc,(uintptr_t)base,len,prot) == 0) {return -EFAULT;}
 
 #define VALIDATE_USER_STRING(proc,str) if (process_is_userspace_string(proc,str) == 0) {return -EFAULT;}
